split promptOverwrite from saveTemplate
