+++ jest.config.js
@@ -1,10 +1,10 @@
 module.exports = {
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
+  testEnvironment: 'node',
   testPathIgnorePatterns: ['./test/e2e'],
   testRegex: '(/test/.*.(test|spec)).(jsx?|tsx?)$',
   testTimeout: 60000,
+  transform: {
+    '^.+\\.tsx?$': 'ts-jest',
+  },
 };
